fin atoi

+++ source/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: klescaud <klescaud@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/01/26 03:36:07 by klescaud          #+#    #+#             */
-/*   Updated: 2015/01/26 03:39:06 by klescaud         ###   ########.fr       */
+/*   Updated: 2015/02/04 21:39:23 by klescaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int		ft_strcmp(const char *s1, const char *s2)
 	int		i;
 
 	i = 0;
-	while (i < ft_strlen(s1))
+	while (i < (int)ft_strlen(s1))
 	{
 		if (s1[i] < s2[i])
 			return (-1);
